chore: extract formatters into own folder
