+++ formatters/index.js
@@ -1,0 +1,6 @@
+const Artifact = require('./artifact');
+const Alias = require('./alias');
+const File = require('./file');
+const Version = require('./version');
+
+module.exports = { Artifact, Alias, Version, File };
